# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_yamovo_contentx_internal_models.ActivityLog:
    properties:
      action:
        description: create, update, delete, login, etc.
        type: string
      created_at:
        type: string
      details:
        type: string
      entity:
        description: article, user, comment, etc.
        type: string
      entity_id:
        type: integer
      id:
        type: integer
      ip:
        type: string
      user_agent:
        type: string
      user_id:
        type: integer
    type: object
...
  github_com_yamovo_contentx_internal_models.Plugin:
    properties:
      author:
        type: string
      config:
        additionalProperties: true
        type: object
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      dependencies:
        description: comma-separated
        type: string
      description:
        type: string
      entry_point:
        type: string
      id:
        type: integer
      is_enabled:
        type: boolean
      min_version:
        description: minimum CMS version
        type: string
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
      version:
        type: string
      website:
        type: string
    type: object
...
  github_com_yamovo_contentx_internal_models.ThemeConfig:
    properties:
      author:
        type: string
      config:
        additionalProperties: true
        type: object
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      screenshot:
        type: string
      slug:
        type: string
      template_dir:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_handlers.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      err_code:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/internal_handlers.Meta'
    type: object
  internal_handlers.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: High-performance Headless CMS - API-first content platform
  title: ContentX API
  version: "1.0"
paths:
...
  /plugins:
    get:
      description: Returns all installed plugins
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_yamovo_contentx_internal_models.Plugin'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: List plugins
      tags:
      - Plugins
  /plugins/{id}/config:
    put:
      consumes:
      - application/json
      description: Updates a plugin's configuration (requires plugins.manage permission)
      parameters:
      - description: Plugin ID
        in: path
        name: id
        required: true
        type: integer
      - description: Plugin configuration
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update plugin config
      tags:
      - Plugins
  /plugins/{id}/disable:
    post:
      description: Disables a plugin (requires plugins.manage permission)
      parameters:
      - description: Plugin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Disable plugin
      tags:
      - Plugins
  /plugins/{id}/enable:
    post:
      description: Enables a plugin (requires plugins.manage permission)
      parameters:
      - description: Plugin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Enable plugin
      tags:
      - Plugins
...
  /system/activity:
    get:
      description: Returns paginated activity audit trail
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Items per page
        in: query
        name: page_size
        type: integer
      - description: Filter by entity type
        in: query
        name: entity
        type: string
      - description: Filter by action
        in: query
        name: action
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_yamovo_contentx_internal_models.ActivityLog'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Activity log
      tags:
      - System
...
  /system/info:
    get:
      description: Returns system information (version, uptime, Go version, etc.)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: System info
      tags:
      - System
...
  /themes:
    get:
      description: Returns all installed themes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handlers.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_yamovo_contentx_internal_models.ThemeConfig'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: List themes
      tags:
      - Themes
  /themes/{id}/activate:
    post:
      description: Activates a theme (requires themes.manage permission)
      parameters:
      - description: Theme ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Activate theme
      tags:
      - Themes
  /themes/{id}/config:
    put:
      consumes:
      - application/json
      description: Updates theme configuration (requires themes.manage permission)
      parameters:
      - description: Theme ID
        in: path
        name: id
        required: true
        type: integer
      - description: Theme configuration
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Update theme config
      tags:
      - Themes
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"